low testss

+++ src/main/java/container/LockFreeSet.java
@@ -62,14 +62,14 @@ public class LockFreeSet<T extends Comparable<T>> implements ILockFreeSet<T> {
     }
 
     @Override
-    public AtomicInteger size() {
+    public int size() {
         Node<T> curr = head.getNext();
         AtomicInteger counter = new AtomicInteger();
         while (curr != null) {
             curr = curr.getNext();
             counter.incrementAndGet();
         }
-        return counter;
+        return counter.intValue();
     }
 
     private Pair<T> search(T value) {
